+++ internal/pkg/http/handlers/api/session/create/handler.go
@@ -5,10 +5,8 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/tarampampam/webhook-tester/internal/pkg/http/responder"
-
 	"github.com/tarampampam/webhook-tester/internal/pkg/http/handlers/api"
-
+	"github.com/tarampampam/webhook-tester/internal/pkg/http/responder"
 	"github.com/tarampampam/webhook-tester/internal/pkg/storage"
 )