+++ internal/pkg/http/handlers/api/session/requests/delete/handler_test.go
@@ -82,12 +82,12 @@ func TestHandler_ServeHTTPSuccess(t *testing.T) {
 	defer func() { _ = ps.Close() }()
 
 	// create session
-	sessionUUID, err := s.CreateSession("foo", 202, "foo/bar", 0)
+	sessionUUID, err := s.CreateSession([]byte("foo"), 202, "foo/bar", 0)
 	assert.NoError(t, err)
 
 	// create 2 requests for the session
-	_, _ = s.CreateRequest(sessionUUID, "", "", "", "", nil)
-	requestUUID, _ := s.CreateRequest(sessionUUID, "", "", "", "", nil)
+	_, _ = s.CreateRequest(sessionUUID, "", "", "", []byte{}, nil)
+	requestUUID, _ := s.CreateRequest(sessionUUID, "", "", "", []byte{}, nil)
 	requests, err := s.GetAllRequests(sessionUUID)
 	assert.NoError(t, err)
 	assert.Len(t, requests, 2) // is not empty