+++ web/js/highlightjs.js
@@ -3,6 +3,19 @@
 /* global define, hljs */
 
 define(['Vue', 'hljs'], (Vue) => {
+    /**
+     * @param {String} code
+     */
+    const formatSourceCode = (code) => {
+        try { // decorate json
+            return JSON.stringify(JSON.parse(code), null, 2);
+        } catch (e) {
+            // wrong json
+        }
+
+        return code; // fallback
+    };
+
     // <https://vuejsfeed.com/blog/vue-js-syntax-highlighting-with-highlight-js>
     Vue.directive('highlightjs', {
         deep: true,
@@ -10,7 +23,7 @@ define(['Vue', 'hljs'], (Vue) => {
             el.querySelectorAll('code').forEach((target) => {
                 if (binding.value) {
                     target.className = 'hljs' // reset highlighting language
-                    target.innerText = binding.value;
+                    target.innerText = formatSourceCode(binding.value);
                 }
                 hljs.highlightElement(target);
             })
@@ -19,7 +32,7 @@ define(['Vue', 'hljs'], (Vue) => {
             el.querySelectorAll('code').forEach((target) => {
                 if (binding.value) {
                     target.className = 'hljs' // reset highlighting language
-                    target.innerText = binding.value;
+                    target.innerText = formatSourceCode(binding.value);
                     hljs.highlightElement(target);
                 }
             })